LPS-170415 Move util files to a more generic package
